move interactive auth flow to public
