+++ apps/internal/base/base.go
@@ -5,9 +5,6 @@ package base
 
 import (
 	"context"
-	"crypto/rand"
-	"crypto/sha256"
-	"encoding/base64"
 	"errors"
 	"fmt"
 	"net/url"
@@ -17,14 +14,11 @@ import (
 	"time"
 
 	"github.com/AzureAD/microsoft-authentication-library-for-go/apps/cache"
-	"github.com/AzureAD/microsoft-authentication-library-for-go/apps/internal/base/internal/local"
 	"github.com/AzureAD/microsoft-authentication-library-for-go/apps/internal/base/internal/storage"
 	"github.com/AzureAD/microsoft-authentication-library-for-go/apps/internal/oauth"
 	"github.com/AzureAD/microsoft-authentication-library-for-go/apps/internal/oauth/ops/accesstokens"
 	"github.com/AzureAD/microsoft-authentication-library-for-go/apps/internal/oauth/ops/authority"
 	"github.com/AzureAD/microsoft-authentication-library-for-go/apps/internal/shared"
-	"github.com/google/uuid"
-	"github.com/pkg/browser"
 )
 
 const (
@@ -64,12 +58,6 @@ type AcquireTokenAuthCodeParameters struct {
 	Challenge  string
 	AppType    accesstokens.AppType
 	Credential *accesstokens.Credential
-}
-
-// AcquireTokenInteractiveParameters contains the parameters required to acquire an access token using
-// the interactive auth code flow.
-type AcquireTokenInteractiveParameters struct {
-	Scopes []string
 }
 
 // AuthResult contains the results of one token acquisition operation in PublicClientApplication
@@ -268,73 +256,6 @@ func (b Client) AcquireTokenByAuthCode(ctx context.Context, authCodeParams Acqui
 	return b.AuthResultFromToken(ctx, authParams, token, true)
 }
 
-// AcquireTokenByInteractive initiates the interactive auth flow.
-// https://docs.microsoft.com/en-us/azure/active-directory/develop/msal-authentication-flows#interactive-and-non-interactive-authentication
-func (b Client) AcquireTokenByInteractive(ctx context.Context, params AcquireTokenInteractiveParameters) (AuthResult, error) {
-	// the code verifier is a random 32-byte sequence that's been base-64 encoded without padding.
-	// it's used to prevent MitM attacks during auth code flow, see https://tools.ietf.org/html/rfc7636
-	cv, hash, err := codeVerifier()
-	if err != nil {
-		return AuthResult{}, err
-	}
-	authParams := b.AuthParams // This is a copy, as we dont' have a pointer receiver and .AuthParams is not a pointer.
-	authParams.Scopes = params.Scopes
-	authParams.AuthorizationType = authority.ATInteractive
-	authParams.CodeChallenge = hash
-	authParams.CodeChallengeMethod = "S256"
-	authParams.State = uuid.New().String()
-	authParams.Prompt = "select_account"
-	res, err := b.browserLogin(ctx, authParams)
-	if err != nil {
-		return AuthResult{}, err
-	}
-	authParams.Redirecturi = res.redirectURI
-
-	req, err := accesstokens.NewCodeChallengeRequest(authParams, accesstokens.ATPublic, nil, res.authCode, cv)
-	if err != nil {
-		return AuthResult{}, err
-	}
-
-	token, err := b.Token.AuthCode(ctx, req)
-	if err != nil {
-		return AuthResult{}, err
-	}
-
-	return b.AuthResultFromToken(ctx, authParams, token, true)
-}
-
-type interactiveAuthResult struct {
-	authCode    string
-	redirectURI string
-}
-
-// browserLogin launches the system browser for interactive login
-func (b Client) browserLogin(ctx context.Context, params authority.AuthParams) (interactiveAuthResult, error) {
-	// start local redirect server so login can call us back
-	srv, err := local.New(params.State)
-	if err != nil {
-		return interactiveAuthResult{}, err
-	}
-	defer srv.Shutdown()
-	authURL, err := b.AuthCodeURL(ctx, params.ClientID, srv.Addr, params.Scopes, params)
-	if err != nil {
-		return interactiveAuthResult{}, err
-	}
-	// open browser window so user can select credentials
-	if err := browser.OpenURL(authURL); err != nil {
-		return interactiveAuthResult{}, err
-	}
-	// now wait until the logic calls us back
-	res := srv.Result(ctx)
-	if res.Err != nil {
-		return interactiveAuthResult{}, res.Err
-	}
-	return interactiveAuthResult{
-		authCode:    res.Code,
-		redirectURI: srv.Addr,
-	}, nil
-}
-
 func (b Client) AuthResultFromToken(ctx context.Context, authParams authority.AuthParams, token accesstokens.TokenResponse, cacheWrite bool) (AuthResult, error) {
 	if !cacheWrite {
 		return NewAuthResult(token, shared.Account{})
@@ -380,16 +301,4 @@ func convertStrUnixToUTCTime(unixTime string) (time.Time, error) {
 		return time.Time{}, err
 	}
 	return time.Unix(timeInt, 0).UTC(), nil
-}
-
-// creates a code verifier string along with its SHA256 hash
-func codeVerifier() (string, string, error) {
-	cvBytes := make([]byte, 32)
-	if _, err := rand.Read(cvBytes); err != nil {
-		return "", "", err
-	}
-	cv := base64.RawURLEncoding.EncodeToString(cvBytes)
-	// for PKCE, create a hash of the code verifier
-	cvh := sha256.Sum256([]byte(cv))
-	return cv, base64.RawURLEncoding.EncodeToString(cvh[:]), nil
 }
